fix: resolve not a directory on mac m1

+++ src/download.rs
@@ -197,9 +197,9 @@ pub fn unpack_ffmpeg(from_archive: &PathBuf, binary_folder: &Path) -> Result<()>
   // Move binaries
   let (ffmpeg, ffplay, ffprobe) = if cfg!(all(target_os = "macos", target_arch = "x86_64")) {
     (
-      temp_folder.join("ffmpeg"),
-      temp_folder.join("ffplay"),  // <-- no ffplay on mac m1
-      temp_folder.join("ffprobe"), // <-- no ffprobe on mac m1
+      temp_folder.join("./ffmpeg"),
+      temp_folder.join("./ffplay"),  // <-- no ffplay on mac m1
+      temp_folder.join("./ffprobe"), // <-- no ffprobe on mac m1
     )
   } else if cfg!(target_os = "windows") {
     let inner_folder = read_dir(&temp_folder)?
@@ -235,7 +235,7 @@ pub fn unpack_ffmpeg(from_archive: &PathBuf, binary_folder: &Path) -> Result<()>
   }
 
   // Delete archive and unpacked files
-  if temp_folder.exists() {
+  if temp_folder.exists() && temp_folder.is_dir() {
     remove_dir_all(&temp_folder)?;
   }
